current work

+++ app/src/main/java/io/github/xlnk/telegramcopy/domain/entity/model/ImColor.kt
@@ -1,4 +1,4 @@
 package io.github.xlnk.telegramcopy.domain.entity.model
 
 @JvmInline
-value class ImColor(val value: ULong)
+value class ImColor(val value: Int)
